✅ TCR - tests passing
changed-lines:
    src: 13
    test: 0
test-stats:
    run: 1
    passed: 1
    failed: 0
    skipped: 0
    error: 0
    duration: 88ms

+++ java/src/main/java/marine/Instruction.java
@@ -3,13 +3,15 @@ package marine;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 public class Instruction {
     public String text;
-    public int x;
+    public BigDecimal x;
 
     public static Instruction fromText(String text) {
         var split = text.split(" ");
-        return new Instruction(split[0], Integer.parseInt(split[1]));
+        return new Instruction(split[0], BigDecimal.valueOf(Integer.parseInt(split[1])));
     }
 }
